pyproject.toml
[project]
name = "aether-env"
version = "0.1.0"
description = "Your project description"
requires-python = "==3.12.*"

dependencies = [
    "google-adk>=1.28.1",
    "fastapi>=0.135.3",
    "uvicorn>=0.41.0",
    "google-genai>=1.70.0",
    "pydantic>=2.12.5",
    "langgraph>=0.4.7",
    "google-auth>=2.49.1",
    "google-auth-oauthlib>=1.3.1",
    "google-api-python-client>=2.193.0",
    "python-dotenv>=1.2.2",
    "fastmcp>=3.2.0",
    "google-adk[extensions]"
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"